Support Swagger-UI validatorUrl option.
When you are running Swagger-UI on a server which is not open to external
access, the default validation behavior causes an error message to be
displayed. This change allows you to set a validator_url option to nil,
which suppresses this error message. You could also point to a validator
service under your own control.